fixes after review: change files structure
