+++ app/src/main/java/an/maguste/android/navier/movieslist/MoviesListViewModelFactory.kt
@@ -1,13 +1,11 @@
-package an.maguste.android.navier.mvvm
+package an.maguste.android.navier.movieslist
 
-import an.maguste.android.navier.data.Movie
 import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 
-// Movies list
 class MoviesListViewModelFactory(private val context: Context,
                                  private val dispatcher: CoroutineDispatcher = Dispatchers.Default
 ) : ViewModelProvider.Factory{
@@ -15,18 +13,7 @@ class MoviesListViewModelFactory(private val context: Context,
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel?> create(modelClass: Class<T>): T = when (modelClass) {
         MoviesListViewModel::class.java -> MoviesListViewModel(context = context,
-                dispatcher = dispatcher)
-        else -> throw IllegalArgumentException("$modelClass is not registered ViewModel")
-    } as T
-
-}
-
-// Movies detail
-class MoviesDetailViewModelFactory(private val movie: Movie) : ViewModelProvider.Factory{
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T = when (modelClass) {
-        MoviesDetailsViewModel::class.java -> MoviesDetailsViewModel(movie,)
+            dispatcher = dispatcher)
         else -> throw IllegalArgumentException("$modelClass is not registered ViewModel")
     } as T
